backend/requirements.txt
fastapi
uvicorn
anthropic
twilio
supabase
python-dotenv
httpx
python-multipart
deep-translator
groq